Backend: implement postgres connection
