+++ .idea/sqldialects.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="SqlDialectMappings">
+    <file url="file://$PROJECT_DIR$/backend/src/main/kotlin/com/github/devinkadrie/basedbin/PasteSchema.kt" dialect="GenericSQL" />
+  </component>
+</project>
