+++ README.md
@@ -1,2 +1,7 @@
 # Basedbin
 
+
+## To Do
+
+- [ ] Use UUID for primary keys
+- [ ] Set up docker compose
