+++ backend/.gitignore
@@ -34,3 +34,4 @@ out/
 
 ### VS Code ###
 .vscode/
+/.kotlin/
