+++ backend/src/main/kotlin/com/github/devinkadrie/basedbin/Application.kt
@@ -8,6 +8,5 @@ fun main(args: Array<String>) = io.ktor.server.netty.EngineMain.main(args)
 fun Application.module() {
     configureHTTP()
     configureSerialization()
-    configureDatabases()
     configureRouting()
 }
