+++ backend/src/main/resources/application.yaml
@@ -4,3 +4,11 @@ ktor:
             - com.github.devinkadrie.basedbin.ApplicationKt.module
     deployment:
         port: 8080
+
+postgres:
+    url:
+        "jdbc:postgresql://localhost:32771/postgres"
+    user:
+        "postgres"
+    password:
+        ""
